# Speed up scope_start