before:
```
test completes_hashmap               ... bench:     722,266 ns/iter (+/- 20,725)
test completes_methods_for_file_open ... bench:     641,672 ns/iter (+/- 22,381)
test completes_methods_for_vec       ... bench: 105,745,576 ns/iter (+/- 4,441,718)
test follows_collections             ... bench:  16,464,915 ns/iter (+/- 448,679)
test follows_std                     ... bench:  21,376,878 ns/iter (+/- 631,305)
```

after:
```
test completes_hashmap               ... bench:     680,692 ns/iter (+/- 30,410)
test completes_methods_for_file_open ... bench:     634,635 ns/iter (+/- 23,081)
test completes_methods_for_vec       ... bench:  93,869,941 ns/iter (+/- 1,277,929)
test follows_collections             ... bench:  15,327,823 ns/iter (+/- 409,412)
test follows_std                     ... bench:  20,758,956 ns/iter (+/- 1,076,299)
```